requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coverage==6.1.2
Django==3.2.9
django-debug-toolbar==3.2.2
django-environ==0.8.1
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pymemcache==3.5.0
python-memcached==1.59
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
